+++ src/main/java/ru/praktika/kotouslugi/dao/LicenceRequisitionRepository.java
@@ -3,5 +3,5 @@ package ru.praktika.kotouslugi.dao;
 import org.springframework.data.repository.CrudRepository;
 import ru.praktika.kotouslugi.model.LicenceRequisition;
 
-public interface LicenceRequisitionRepository extends CrudRepository<LicenceRequisition, Integer> {
+public interface LicenceRequisitionRepository extends CrudRepository<LicenceRequisition, Long> {
 }
